Fixing silent-input over password
--silent-input was being ignored on password. This is a little fix
I have also moved an abstract just for InputAbstract
